Rename scenarios
