+++ cypress/e2e/internal/charge-information/sroc/validate-new-charge-infromation.cy.js
